// src/app/interfaces/user.ts
export interface User {
    avatar?: any | null;
    bgImage?: any | null;
    birthday?: any;
    city?: string;
    country?: string;
    createdDate?: number;
    displayName?: string;
    id?: string;
    language?: string;
    lastActiveDate?: number;
    level?: number;
    name?: object;
    online?: boolean;
    privateProfile?: boolean;
    showAge?: boolean;
    status?: string;
    title?: string;
    totalPoints?: number;
}